Swift 1.2

+++ Application/Color.swift
@@ -14,7 +14,7 @@ struct Color {
         let red = Double((hex & 0xFF0000) >> 16) / 255.0
         let green = Double((hex & 0x00FF00) >> 8) / 255.0
         let blue = Double(hex & 0x0000FF) / 255.0
-        return UIColor(red: red, green: green, blue: blue, alpha: alpha)
+        return UIColor(red: CGFloat(red), green: CGFloat(green), blue: CGFloat(blue), alpha: alpha)
     }
     
     static func grayDarkest() -> UIColor {
